Update/correction to Saskatchewan, Canada COVID-19 cases and deaths from 6 Feb 2022
I have updated confirmed cases and deaths according to the weekly data published at this location: https://publications.saskatchewan.ca/#/categories/5688
These reports are for weeks ending on Saturdays. I have used the data in the reports for the following week (Sunday-Saturday). For example, the data from the report for Feb 6-12 is recorded in dates Feb 13-19.